move config.json to root of repo
