+++ src/services/juxt-web/routes/admin/api.js
@@ -2,7 +2,7 @@ var express = require('express');
 const database = require('../../../../database');
 const logger = require('../../../../logger');
 const util = require('../../../../util');
-const config = require('../../../../config.json');
+const config = require('../../../../../config.json');
 const { COMMUNITY } = require('../../../../models/communities');
 const { POST } = require('../../../../models/post');
 var router = express.Router();
