+++ src/services/juxt-web/routes/admin/home.js
@@ -3,7 +3,7 @@ var xml = require('object-to-xml');
 const database = require('../../../../database');
 const logger = require('../../../../logger');
 const util = require('../../../../util');
-const config = require('../../../../config.json');
+const config = require('../../../../../config.json');
 const request = require("request");
 var path = require('path');
 var moment = require('moment');
