+++ src/services/juxt-web/routes/console/messages.js
@@ -2,7 +2,7 @@ var express = require('express');
 var xml = require('object-to-xml');
 const database = require('../../../../database');
 const util = require('../../../../util');
-const config = require('../../../../config.json');
+const config = require('../../../../../config.json');
 const { CONVERSATION } = require('../../../../models/conversation');
 const { POST } = require('../../../../models/post');
 var moment = require('moment');
