+++ src/util.js
@@ -3,7 +3,7 @@ const NodeRSA = require('node-rsa');
 const fs = require('fs-extra');
 const database = require('./database');
 const logger = require('./logger');
-const config = require('./config.json');
+const config = require('../config.json');
 const { USER } = require('./models/user');
 const translations = require('./translations')
 var HashMap = require('hashmap');
